Some tests for the ffmpeg code.
This required some clean up to the handling of default converter options.

Whether it's a shallow copy containing shared never changing slice pointers, or it's a pointer to defaults that should never be changed, it's kinda a wort.

Since Go lacks const references, I say six of one, half of the other, tests are better.