Exporter is now asyncronous.
This separates the fs.WalkDir into two separate walks:

1. Creating each directory.
2. Copying and converting each file.

With the second walk functioning by adding the operations to a work pool.

Further, the ffmpeg run is now performed in the background and sent the the log (if any).

In general, this takes the test case of converting my library to FLAC from a runtime of 30-40 minutes down to 5-10 minutes when running on my MacBook Air 2022 (M2 / 16 GB). Consumes most of the RAM, but running 8 concurrent flac conversions is pretty darn quick and actually takes advantage of these cores :).