Major reorg of options.
This effectively splits the various options into following the same type of
pattern, and chunks out the flag set and validation code into their own methods.
Such that it's easy to have one flag set with options from multiple embedded
types. Perhaps in the future, this will get cleaned up into a nice interface.

For now, the main purpose is to be able to have ExporterOptions embed
ConverterOptions, so that it's possible for export_audio_tree to override the
default ConverterOptions in the same manner as the various to_* programs.